Improved Dweller path finding - Auto retargeting system (if there are given default targets); allows dwellers to traverse a maze periodically
